==> 0_paths.R <==

# path_to_box <- "C:/Users/Natalie Gradwohl/Box/"

# path_to_box <- "C:/Users/josephfong/Box/"
# path_to_box <- "/Users/yingyanwu/Library/CloudStorage/Box-Box/"
# path_to_box <- c("C:/Users/lrojas-saunero/Box/")
# path_to_box <- "C:/Users/Staff/Box/"
path_to_box <- "~/Library/CloudStorage/Box-Box/"

path_to_raw_data <- paste0(path_to_box,
                           "Asian_Americans_dementia_data/analysis_data_tables/")

path_to_analytic_data <- paste0(path_to_box,
                                "Asian_Americans_dementia_data/bp_dementia/")

path_to_bp_data <- paste0(path_to_box,
                          "Asian_Americans_dementia_data/raw_data_tables/Membership_EHR_mortality_data/")
